requirements.txt
bitfinex-v2
matplotlib
numpy
pandas
PyYAML
statsmodels
